Bundle Javadoc with Antora
- Remove javadoc related tasks from root project
- Add aggregated javadoc task to spring-shell-docs
- Various antora version updates
- Bundle javadoc via collector into antora build
- Remove refs to distZip in workflows
- Relates #1132
