+++ .github/workflows/release-ga.yml
@@ -59,7 +59,7 @@ jobs:
         GRADLE_ENTERPRISE_CACHE_PASSWORD: ${{ secrets.GRADLE_ENTERPRISE_CACHE_PASSWORD }}
         GRADLE_ENTERPRISE_ACCESS_KEY: ${{ secrets.GRADLE_ENTERPRISE_SECRET_ACCESS_KEY }}
       run: |
-        jf rt gradle build distZip artifactoryPublish
+        jf rt gradle build artifactoryPublish
         jf rt build-publish
     - name: Push Release
       env:
