+++ build.gradle
@@ -1,6 +1,5 @@
 plugins {
 	id "base"
-	id 'org.springframework.shell.root'
 }
 
 description = 'Spring Shell'
